Create the optimization variable :hydro_spill for hydro storage nodes.
This enables reservoirs to spill excess water without producing energy.
Remove the check that denied creating a hydro storage node with too much
inflow, and replace it with a check that only forces the initial level
to be less than the maximum storage level. Create a test for the new
hydro storage variable and update tests with the :hydro_spill variable.